Move results fragment to card component
